Move TimeSettings into the absence package
