+++ src/test/java/org/synyx/urlaubsverwaltung/absence/AbsenceTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Test;
 import org.synyx.urlaubsverwaltung.period.DayLength;
 import org.synyx.urlaubsverwaltung.period.Period;
 import org.synyx.urlaubsverwaltung.person.Person;
-import org.synyx.urlaubsverwaltung.settings.TimeSettings;
 
 import java.time.Clock;
 import java.time.LocalDate;
