+++ src/test/java/org/synyx/urlaubsverwaltung/calendar/DepartmentCalendarServiceTest.java
@@ -15,7 +15,7 @@ import org.synyx.urlaubsverwaltung.period.DayLength;
 import org.synyx.urlaubsverwaltung.period.Period;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
-import org.synyx.urlaubsverwaltung.settings.TimeSettings;
+import org.synyx.urlaubsverwaltung.absence.TimeSettings;
 
 import java.io.File;
 import java.time.LocalDate;
