+++ src/test/java/org/synyx/urlaubsverwaltung/calendarintegration/CalendarMailServiceTest.java
@@ -12,7 +12,7 @@ import org.synyx.urlaubsverwaltung.mail.Mail;
 import org.synyx.urlaubsverwaltung.mail.MailService;
 import org.synyx.urlaubsverwaltung.period.Period;
 import org.synyx.urlaubsverwaltung.person.Person;
-import org.synyx.urlaubsverwaltung.settings.TimeSettings;
+import org.synyx.urlaubsverwaltung.absence.TimeSettings;
 
 import java.time.LocalDate;
 import java.util.HashMap;
